+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0] - 2020-10-13
 ### Added
 - Log tags now comes with `ISO_TIMESTAMP` which produce timestamp such as `2020-10-13T08:46:38.803Z`.
 - Default timestamp formatter and default message formatter are exported for use.
@@ -40,7 +42,8 @@ and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 
 
-[Unreleased]: https://github.com/VeryCrazyDog/level-logger/compare/0.1.2...HEAD
+[Unreleased]: https://github.com/VeryCrazyDog/level-logger/compare/1.0.0...HEAD
+[1.0.0]: https://github.com/VeryCrazyDog/level-logger/compare/0.1.2...1.0.0
 [0.1.2]: https://github.com/VeryCrazyDog/level-logger/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/VeryCrazyDog/level-logger/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/VeryCrazyDog/level-logger/releases/tag/0.1.0